==> station_positioning/2_writeDfTest.ipynb <==
# Preparazione data set per i test
Loading dei dati necessari ai test e suddivisione in 3 Dataframe: 
* dfLow = zona a bassa coverage;
* dfNormal = zona a media coverage
* dfHigh = zona ad alta coverage

#scrittura df per test
import pandas as pd
from shapely.geometry import Point, Polygon,box
import yaml
import geopandas as gpd
#import cartopy.geodesic as GD
import numpy as np
from datetime import datetime
import os


with open("../conf.yaml") as f:
    conf = yaml.load(f, Loader = yaml.FullLoader)

pathBox = conf["pathBox"]

# Check and read the start end period period for computing the coverage. 
if "coverage_start_period" and "coverage_end_period" in conf:
    coverage_start_period = conf["coverage_start_period"]
    coverage_end_period = conf["coverage_end_period"]
else:
    # coverage period overlaps with the start and end time of the data set
    coverage_start_period = conf["start_time"]
    coverage_end_period = conf["end_time"]

start_time_obj = datetime.strptime(coverage_start_period, '%Y-%m-%d %H:%M:%S')
end_time_obj = datetime.strptime(coverage_end_period, '%Y-%m-%d %H:%M:%S')
a = str(start_time_obj.year)+"_"+str(start_time_obj.month)+"_"+str(start_time_obj.day)
b = str(end_time_obj.year)+"_"+str(end_time_obj.month)+"_"+str(end_time_obj.day)
# build suffix for the output file namnes
suffix = a+"-"+b
out_path = conf["out_path"]

coverage_path = "../"+out_path+ "scenario/mix/"+str(conf["scale_value"])+"/coverages_"+suffix+".csv"
coverage_path





# build the folders 
def ensure_dir(file_path):
    directory = os.path.dirname(file_path)
    if not os.path.exists(directory):
        os.makedirs(directory)
        
# build output paths
pathDFLow = conf["pathDFLow"]+suffix+"/"
pathDFNormal = conf["pathDFNormal"]+suffix+"/"
pathDFHigh = conf["pathDFHigh"]+suffix+"/"
pathDFBorder = conf["pathDFBorder"]+suffix+"/"

ensure_dir(pathDFLow)
ensure_dir(pathDFNormal)
ensure_dir(pathDFHigh)
ensure_dir(pathDFBorder)



#read the coverage file
GeneralDf = pd.read_csv(coverage_path, delimiter=',')
#rinomino le colonne
#GeneralDf.rename(columns={'Unnamed: 0': 'index', 'Unnamed: 0.1' : 'index2'}, inplace=True)
#elimino le colonne che non servono
#GeneralDf.drop(GeneralDf.columns[[0,1,2,3,6]], axis='columns', inplace=True)
print(GeneralDf.head())

#trasformo il dataframe in un gedoDataFrame
gdf = gpd.GeoDataFrame(GeneralDf,crs="EPSG:4326", geometry=gpd.points_from_xy(GeneralDf.lon, GeneralDf.lat))
gdf = gdf.drop(['lon', 'lat'], axis=1)
print(gdf.head())

#lettura dei box
Box = gpd.GeoDataFrame(gpd.read_file(pathBox),crs="EPSG:4326")
Box=Box.set_index('index')
print(Box)


boxLow=Box.at["boxLow",'geometry']
pdfLow = gpd.GeoDataFrame({'geometry':[boxLow]},crs="EPSG:4326")
boxNormal=Box.at["boxNormal",'geometry']
pdfNormal = gpd.GeoDataFrame({'geometry':[boxNormal]},crs="EPSG:4326")
boxHigh=Box.at["boxHigh",'geometry']
pdfHigh = gpd.GeoDataFrame({'geometry':[boxHigh]},crs="EPSG:4326")
boxBorder=Box.at["boxBorder",'geometry']
pdBorder = gpd.GeoDataFrame({'geometry':[boxBorder]},crs="EPSG:4326")

#SJOIN
#high
dfHigh = gpd.sjoin(gdf, pdfHigh, how = "inner", op="intersects")
dfHigh = dfHigh.drop(['index_right'],axis=1)
if dfHigh.empty == False:
    dfHigh.to_file(pathDFHigh+"DFHigh_"+suffix+".geojson",driver='GeoJSON')
print("Locazioni all'interno del Box High:",len(dfHigh))

#low
dfLow = gpd.sjoin(gdf, pdfLow, how = "inner", op="intersects")
dfLow = dfLow.drop(['index_right'],axis=1)
if dfLow.empty == False:
    dfLow.to_file(pathDFLow+"DFLow_"+suffix+".geojson",driver='GeoJSON')
print("Locazioni all'interno del Box Low:", len(dfLow))

#normal
dfNormal = gpd.sjoin(gdf, pdfNormal, how = "inner", op="intersects")
dfNormal = dfNormal.drop(['index_right'],axis=1)
if dfNormal.empty == False:
    dfNormal.to_file(pathDFNormal+"DFNormal_"+suffix+".geojson",driver='GeoJSON')
print("Locazioni all'interno del Box Normal:",len(dfNormal))

#Border
dfBorder = gpd.sjoin(gdf, pdBorder, how = "inner", op="intersects")
dfBorder = dfBorder.drop(['index_right'],axis=1)
if dfBorder.empty == False:
    dfBorder.to_file(pathDFBorder+"DFBorder_"+suffix+".geojson",driver='GeoJSON')
print("Locazioni all'interno del Box Boder:",len(dfBorder))



# Opzionale
# ampliamento del dfLow con posizione e probabilità generate in maniera casuale e salvataggio su file

#aggiunta dfLow
boundsBox = boxLow.bounds
minlat=boundsBox[1]
maxlat=boundsBox[3]
minlon=boundsBox[0]
maxlon=boundsBox[2]

probability = np.random.uniform(0,0.6,250)
xpoints = np.random.uniform(0,0.1,250) + minlon
ypoints = np.random.uniform(0,0.1,250)+minlat
d = {'probability':probability,'geometry': gpd.points_from_xy(xpoints,ypoints)}
df = gpd.GeoDataFrame(d,crs="EPSG:4326")
dfLow=dfLow.append(df)
dfLow.to_file(pathDFLow+"DFLow.geojson",driver='GeoJSON')
dfLow


# ampliamento del dfNormal con posizione e probabilità generate in maniera casuale e salvataggio su file

#aggiunta dfNormal
boundsBox = boxNormal.bounds
minlat=boundsBox[1]
maxlat=boundsBox[3]
minlon=boundsBox[0]
maxlon=boundsBox[2]

probability = np.random.uniform(0,1,200)
xpoints = np.random.uniform(0,0.1,200) + minlon
ypoints = np.random.uniform(0,0.1,200)+minlat
d = {'probability':probability,'geometry': gpd.points_from_xy(xpoints,ypoints)}
df = gpd.GeoDataFrame(d,crs="EPSG:4326")
dfNormal=dfNormal.append(df)
dfNormal.to_file(pathDFNormal+"DFNormal.geojson",driver='GeoJSON')
dfNormal

# ampliamento del dfHigh con posizione e probabilità generate in maniera casuale e salvataggio su file

#aggiunta dfHigh
boundsBox = boxHigh.bounds
minlat=boundsBox[1]
maxlat=boundsBox[3]
minlon=boundsBox[0]
maxlon=boundsBox[2]

augemented_locations = 1000
probability = np.random.uniform(0,0.6,augemented_locations)
xpoints = np.random.uniform(0,0.1,augemented_locations) + minlon
ypoints = np.random.uniform(0,0.1,augemented_locations)+minlat
id_locations = np.random.uniform(1,100000,augemented_locations)

d = {'probability':probability,'geometry': gpd.points_from_xy(xpoints,ypoints),'id_location':id_locations}
df = gpd.GeoDataFrame(d,crs="EPSG:4326")
dfHigh=dfHigh.append(df)
dfHigh.to_file(pathDFHigh+"DFHigh.geojson",driver='GeoJSON')
dfHigh
