==> station_positioning/6_Bechmark_Tests.ipynb <==
# Posizionamento con altri algoritmi
- 1. Random
- 2. DBSCAN

import pandas as pd
import shapely
#import pyproj
#from pyproj import CRS
from shapely.geometry import Point, Polygon
import geopandas
import matplotlib
#from planar import BoundingBox
#from geopy import distance 
import matplotlib.pyplot as plt
from matplotlib import cm
from scipy.interpolate import griddata
import numpy as np
#import planar
import yaml
import matplotlib.patches as mpatches
from matplotlib.collections import PatchCollection
import cartopy.geodesic as GD


with open('../conf.yaml') as f:
    conf = yaml.load(f, Loader=yaml.FullLoader)
path = conf["path"]
Dr = conf["radius"]
Cr = conf["soglia"]
R0_iterations = conf["R0_iterations"]
pathAreeProibite = conf["pathAreeProibite"]
pathDFLow = conf["pathDFLow"]
pathDFNormal = conf["pathDFNormal"]
pathDFHigh = conf["pathDFHigh"]
pathDFBorder = conf ["pathDFBorder"]
pathDFFull = conf ["pathDFFull"]
pathBox = conf["pathBox"]

#lettura dei df aumentati
# Low
b = geopandas.read_file(pathDFLow+"DFLow.geojson",index_col = 0,columns=['probability','geometry'])
dfLow=geopandas.GeoDataFrame(b,crs="EPSG:4326")
dfLowDiscovered = dfLow[(dfLow.probability<=Cr)]

# Normal
b = geopandas.read_file(pathDFNormal+"DFNormal.geojson",index_col = 0,columns=['probability','geometry'])
dfNormal = geopandas.GeoDataFrame(b,crs="EPSG:4326")
dfNormalDiscovered = dfNormal[(dfNormal.probability<=Cr)]


# High
b = geopandas.read_file(pathDFHigh+"DFHigh.geojson",index_col = 0,columns=['probability','geometry'])
dfHigh=geopandas.GeoDataFrame(b,crs="EPSG:4326")
dfHighDiscovered = dfHigh[(dfHigh.probability<=Cr)]

# Border
b = geopandas.read_file(pathDFBorder+"DFBorder.geojson",index_col = 0,columns=['probability','geometry'])
dfBorder=geopandas.GeoDataFrame(b,crs="EPSG:4326")
dfBorderDiscovered = dfBorder[(dfBorder.probability<=Cr)]


#prelevo i box per il test
Box = geopandas.GeoDataFrame(geopandas.read_file(pathBox),crs="EPSG:4326")
Box=Box.set_index('index')
boxLow=Box.at["boxLow",'geometry']
boxNormal=Box.at["boxNormal",'geometry']
boxHigh=Box.at["boxHigh",'geometry']
boxBorder=Box.at["boxBorder",'geometry']


#Prelevo le aree proibite
areeProibite = geopandas.GeoDataFrame(geopandas.read_file(pathAreeProibite),crs = "EPSG:4326")
areeProibite=areeProibite['geometry']


# 1. Random Deployement

# Funzione per il posizionamento random di una stazione


def randomTest1(dfPoint,dfPointDiscovered,box) :

    boundsBox = box.bounds
    minlat=boundsBox[1]
    maxlat=boundsBox[3]
    minlon=boundsBox[0]
    maxlon=boundsBox[2]
    mat = np.zeros((R0_iterations,2))

    #genero lon e lat casuali delle stazioni
    for k in range(R0_iterations) : 
        lon = np.random.uniform (minlon,maxlon,1)
        mat[k][0] = lon
        lat = np.random.uniform (minlat,maxlat,1)
        mat [k][1] = lat
        
    #costruisco un geodataframe con le posizioni generate casualmente
    dfStations = geopandas.GeoDataFrame(crs="EPSG:4326", geometry=geopandas.points_from_xy(mat[:,0], mat[:,1]))
    dfStations['pointsCoveredUAV'] = ""
    dfStations['pointsCoveredTotal'] = ""
    
    #conteggio i punti coperti dalle varie stazioni
    for row in dfStations.itertuples():
        lon = row.geometry.x
        lat = row.geometry.y
        circle_points = GD.Geodesic().circle( lon= lon, lat=lat, radius = Dr)
        circle_poly = Polygon(circle_points)
        dfCircle = geopandas.GeoDataFrame({'geometry':[circle_poly]},crs="EPSG:4326")
        df = geopandas.sjoin(dfPointDiscovered, dfCircle, how = "inner", op="intersects")
        dfStations.at[row.Index,'pointsCoveredUAV'] = df.shape[0]
        df = geopandas.sjoin(dfPoint, dfCircle, how = "inner", op="intersects")
        dfStations.at[row.Index,'pointsCoveredTotal'] = df.shape[0]
    
    print(dfStations.head())
    return dfStations


# Funzione per il posizionamento random di 3 stazioni

import cartopy.geodesic as GD

def randomTest3(dfPoint,dfPointDiscovered,box) :

    boundsBox = box.bounds
    minlat=boundsBox[1]
    maxlat=boundsBox[3]
    minlon=boundsBox[0]
    maxlon=boundsBox[2]
    mat = np.zeros((R0_iterations,6))

    #genero lon e lat casuali delle stazioni
    for k in range(R0_iterations) : 
        for i in range (3) :
            lon = np.random.uniform (minlon,maxlon,1)
            mat[k][0+i*2] = lon
            lat = np.random.uniform (minlat,maxlat,1)
            mat [k][1+i*2] = lat

    #costruisco un geodataframe con le posizioni generate casualmente
    dfStations = pd.DataFrame(mat,columns=['lon1','lat1','lon2','lat2','lon3','lat3'])
    dfStations['pointsCoveredUAV'] = ""
    dfStations['pointsCoveredTotal'] = ""

    #conteggio i punti coperti dalle varie stazioni
    for row in dfStations.itertuples():
        circle_poly1 = Polygon (GD.Geodesic().circle( lon= row.lon1, lat=row.lat1, radius = Dr))
        circle_poly2 = Polygon(GD.Geodesic().circle( lon= row.lon2, lat=row.lat2, radius = Dr))
        circle_poly3 = Polygon(GD.Geodesic().circle( lon= row.lon3, lat=row.lat3, radius = Dr))
        dfCircle = geopandas.GeoDataFrame({'geometry':[circle_poly1,circle_poly2,circle_poly3]},crs="EPSG:4326")
        
        df = geopandas.sjoin(dfPointDiscovered, dfCircle, how = "inner", op="intersects")
        df.drop_duplicates(subset=['geometry'],inplace = True)
        dfStations.at[row.Index,'pointsCoveredUAV'] = df.shape[0]
        df = geopandas.sjoin(dfPoint, dfCircle, how = "inner", op="intersects")
        df.drop_duplicates(subset=['geometry'],inplace = True)
        dfStations.at[row.Index,'pointsCoveredTotal'] = df.shape[0]
    print(dfStations.head())
    return dfStations

# Funzione per il posizionamento random di 3 stazioni con zone non ammesse

def randomTest3_areeProibite(dfPoint,dfPointDiscovered,box) :

    boundsBox = box.bounds
    minlat=boundsBox[1]
    maxlat=boundsBox[3]
    minlon=boundsBox[0]
    maxlon=boundsBox[2]
    mat = np.zeros((R0_iterations,6))

    #genero lon e lat casuali delle stazioni e controllo che siano valide
    for k in range(R0_iterations) : 
        i = 0
        while i<3 :
            lon = np.random.uniform (minlon,maxlon,1)
            lat = np.random.uniform (minlat,maxlat,1)
            if invalidCenter((lon,lat),areeProibite) == False:
                mat[k][0+i*2] = lon
                mat [k][1+i*2] = lat
                i = i+1

    #costruisco un dataframe con le posizioni generate casualmente
    dfStations = pd.DataFrame(mat,columns=['lon1','lat1','lon2','lat2','lon3','lat3'])
    dfStations['pointsCoveredUAV'] = ""
    dfStations['pointsCoveredTotal'] = ""

    #conteggio i punti coperti dalle varie stazioni
    for row in dfStations.itertuples():
        circle_poly1 = Polygon (GD.Geodesic().circle( lon= row.lon1, lat=row.lat1, radius = Dr))
        circle_poly2 = Polygon(GD.Geodesic().circle( lon= row.lon2, lat=row.lat2, radius = Dr))
        circle_poly3 = Polygon(GD.Geodesic().circle( lon= row.lon3, lat=row.lat3, radius = Dr))
        dfCircle = geopandas.GeoDataFrame({'geometry':[circle_poly1,circle_poly2,circle_poly3]},crs="EPSG:4326")
        
        df = geopandas.sjoin(dfPointDiscovered, dfCircle, how = "inner", op="intersects")
        df.drop_duplicates(subset=['geometry'],inplace = True)
        dfStations.at[row.Index,'pointsCoveredUAV'] = df.shape[0]
        df = geopandas.sjoin(dfPoint, dfCircle, how = "inner", op="intersects")
        df.drop_duplicates(subset=['geometry'],inplace = True)
        dfStations.at[row.Index,'pointsCoveredTotal'] = df.shape[0]
    print(dfStations.head())
    return dfStations

# Calcolo di dati statistici e scrittura su file


def statistics(dfStations,path,k) :

    pointsCoveredUAV = dfStations['pointsCoveredUAV']
    pointsCoveredTotal = dfStations['pointsCoveredTotal']
    meanCoverageUAV = pointsCoveredUAV.mean(axis=0)
    stdCoverageUAV = pointsCoveredUAV.std()
    percentile_75UAV = pointsCoveredUAV.quantile(.75)
    percentile_25UAV = pointsCoveredUAV.quantile(.25)
    maxUAV = pointsCoveredUAV.max()
    meanCoverageTot = pointsCoveredTotal.mean(axis=0)
    percentile_75Tot = pointsCoveredTotal.quantile(.75)
    percentile_25Tot = pointsCoveredTotal.quantile(.25)
    maxTot = pointsCoveredTotal.max()
    toWrite = "mean Coverage UAV = " + str(meanCoverageUAV) +"std Coverage UAV ="+str(stdCoverageUAV)+ " 75 Percentile UAV = " + str(percentile_75UAV) + " 25 Percentile UAV= " + str(percentile_25UAV) + " maxUAV="+str(maxUAV)+"\n mean Coverage Tot = " + str(meanCoverageTot) + " 75 Percentile Tot = " + str(percentile_75Tot) + " 25 Percentile Tot= " + str(percentile_25Tot) +" maxTot = " + str(maxTot)
    file = open(path+"randomTest"+str(k) + ".odt","w")
    file.write(toWrite)
    file.close()

    

# Test 
## posizionamento random di 1 stazione in 3 scenari di coverage

#dfStationsLow = randomTest1(dfLow,dfLowDiscovered,boxLow)
#dfStationsNormal = randomTest1(dfNormal,dfNormalDiscovered,boxNormal)
dfStationsHigh = randomTest1(dfHigh,dfHighDiscovered,boxHigh)

#statistics (dfStationsLow,pathDFLow,1)
#statistics(dfStationsNormal,path  DFNormal,1)
statistics(dfStationsHigh,pathDFHigh,1)

# Test 
## posizionamento random di 3 stazioni in 3 scenari di coverage

dfStationsLow = randomTest3(dfLow,dfLowDiscovered,boxLow)
dfStationsNormal = randomTest3(dfNormal,dfNormalDiscovered,boxNormal)
dfStationsHigh = randomTest3(dfHigh,dfHighDiscovered,boxHigh)

statistics (dfStationsLow,pathDFLow,3)
statistics(dfStationsNormal,pathDFNormal,3)
statistics(dfStationsHigh,pathDFHigh,3)

# Test
## posizionamento random di 3 stazioni + zone non ammesse

dfStationsLow = randomTest3_areeProibite(dfLow,dfLowDiscovered,boxLow)
dfStationsNormal = randomTest3_areeProibite(dfNormal,dfNormalDiscovered,boxNormal)
dfStationsHigh = randomTest3_areeProibite(dfHigh,dfHighDiscovered,boxHigh)

statistics (dfStationsLow,pathDFLow,3.1)
statistics(dfStationsNormal,pathDFNormal,3.1)
statistics(dfStationsHigh,pathDFHigh,3.1)

# 2. DBSCAN Deployment
